fix cast

+++ tasks/tbb/sarafanov_m_CanonMatMul_tbb/src/CanonMatrix.cpp
@@ -88,7 +88,7 @@ void CanonMatrix::operator+=(const CanonMatrix& canon_matrix) {
     oneapi::tbb::parallel_for(oneapi::tbb::blocked_range<size_t>(0, size_, size_ / ppc::util::GetPPCNumThreads()),
                               [&](const oneapi::tbb::blocked_range<size_t>& distance) {
                                 for (size_t i = distance.begin(); i != distance.end(); ++i) {
-                                  for (size_t j = 0; j != static_cast<int>(size_); ++j) {
+                                  for (size_t j = 0; j != size_; ++j) {
                                     matrix_[(i * size_) + j] += b_matrix[(j * size_) + i];
                                   }
                                 }
